Add option to filter list of TeamMembers by Team name
This commit adds all of the remaining team members of Product/Engineering and
matches them to their team name.

It also introduces the functionality to filter the view of team members by team
name itself. This preference is then also saved to local storage which allows
the saving of this particular preference.