Use fancy grid sizing to automatically determine column count
`repeat(auto-fit, <size>)` will repeat the `size` however many times
it can before the columns overflow. We can use this because the grid
has a determined width.
Note that there is a slight difference between `auto-fit` and
`auto-fill` but I don't know what it is and it seems to not matter in
this case.

The size for each column is `minmax(27rem, 1fr)`. This means that the
minimum width of each column would be `27rem`, and the maximum would
be `1fr`, meaning all columns will share the same width.

What this means is that the rendering engine will try to fit as many
columns (with a width of `27rem`) as it can. then it will distribute
the width equally between these columns. This is neat.