Unregister Service Worker
This commit aims to prevent caching issues where users would not initially see
the most up-to-date version of the website and would instead have to force
refresh in order to see the latest change.

It turns out, that this is because of the registration of a service worker and
so the browser preserves the cached version of the site when it is updated.

Since this is not needed at the moment, we can safely disable it.