Improve overall performance
This commit implements a few performance improvements mainly aimed around the
app handles each TeamMembers current time and displaying it in a performant way.

By moving the current time (and Momment.js object) into a React Context and
removing the repeating of the Moment object for each team member -- it allows
the app to show the current time in a more performant way.

Plus the handling of the Moment object is now much easier to manage and demands
much less processing power from the client to update every second.